+++ zebedee-reader/pom.xml
@@ -33,7 +33,7 @@
     <dependencies>
         <!--Restolino-->
         <dependency>
-            <groupId>com.github.davidcarboni</groupId>
+            <groupId>com.github.ONSdigital</groupId>
             <artifactId>restolino</artifactId>
         </dependency>
